# docs/marshaller_swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Marshaller API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/characteristic-paths/{serviceId}/{characteristicId}": {
            "get": {
                "description": "returns the content-variable paths of the service that carry the given characteristic",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characteristic-paths"
                ],
                "summary": "list the paths a characteristic can be read from",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the service",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the characteristic",
                        "name": "characteristicId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/marshaller.CharacteristicsPathResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "the service is unknown",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/configurables": {
            "get": {
                "description": "returns the values that can be configured alongside the given characteristic in the listed services",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configurables"
                ],
                "summary": "find configurables for a characteristic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the characteristic that is already set",
                        "name": "characteristicId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "comma separated service ids",
                        "name": "serviceIds",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurables.Configurable"
                            }
                        }
                    },
                    "400": {
                        "description": "characteristicId or serviceIds is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "description": "the request-body form of GET /configurables, for callers that hold the services already",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "configurables"
                ],
                "summary": "find configurables for a characteristic",
                "parameters": [
                    {
                        "description": "the characteristic and the services to search",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.FindConfigurablesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurables.Configurable"
                            }
                        }
                    },
                    "400": {
                        "description": "the body is not valid json, or characteristic_id is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/converter/extension-call": {
            "post": {
                "description": "runs a single converter extension against an input and returns its output, for developing an extension",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "converter"
                ],
                "summary": "try a converter extension",
                "parameters": [
                    {
                        "description": "the extension and the input to run it on",
                        "name": "call",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/converter.ExtensionCall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/converter.ExtensionCallResponse"
                        }
                    },
                    "400": {
                        "description": "the body is not valid json, or the extension failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "the service runs without a converter",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "configurables.Configurable": {
            "type": "object",
            "properties": {
                "characteristic_id": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurables.ConfigurableCharacteristicValue"
                    }
                }
            }
        },
        "configurables.ConfigurableCharacteristicValue": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "converter.ExtensionCall": {
            "type": "object",
            "properties": {
                "extension": {
                    "$ref": "#/definitions/models.ConverterExtension"
                },
                "input": {}
            }
        },
        "converter.ExtensionCallResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "output": {}
            }
        },
        "marshaller.CharacteristicsPathResponse": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "service_characteristic_id": {
                    "type": "string"
                }
            }
        },
...
        "messages.FindConfigurablesRequest": {
            "type": "object",
            "properties": {
                "characteristic_id": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Service"
                    }
                }
            }
        },
...
        "model.Service": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attribute"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "interaction": {
                    "$ref": "#/definitions/models.Interaction"
                },
                "local_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "protocol_id": {
                    "type": "string"
                },
                "service_group_key": {
                    "type": "string"
                }
            }
        },
        "models.Attribute": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Content": {
            "type": "object",
            "properties": {
                "content_variable": {
                    "$ref": "#/definitions/models.ContentVariable"
                },
                "id": {
                    "type": "string"
                },
                "protocol_segment_id": {
                    "type": "string"
                },
                "serialization": {
                    "$ref": "#/definitions/models.Serialization"
                }
            }
        },
        "models.ContentVariable": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "description": "deprecated: please use AspectIds",
                    "type": "string"
                },
                "aspect_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "characteristic_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_void": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "omit_empty": {
                    "type": "boolean"
                },
                "serialization_options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sub_content_variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContentVariable"
                    }
                },
                "type": {
                    "$ref": "#/definitions/models.Type"
                },
                "unit_reference": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "models.ConverterExtension": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "integer"
                },
                "formula": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "placeholder_name": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "models.Interaction": {
            "type": "string",
            "enum": [
                "event",
                "request",
                "event+request"
            ],
            "x-enum-varnames": [
                "EVENT",
                "REQUEST",
                "EVENT_AND_REQUEST"
            ]
        },
...
        "models.Serialization": {
            "type": "string",
            "enum": [
                "xml",
                "json",
                "plain-text"
            ],
            "x-enum-varnames": [
                "XML",
                "JSON",
                "PlainText"
            ]
        },
        "models.Type": {
            "type": "string",
            "enum": [
                "https://schema.org/Text",
                "https://schema.org/Integer",
                "https://schema.org/Float",
                "https://schema.org/Boolean",
                "https://schema.org/ItemList",
                "https://schema.org/StructuredValue"
            ],
            "x-enum-varnames": [
                "String",
                "Integer",
                "Float",
                "Boolean",
                "List",
                "Structure"
            ]
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}